refactor: split the automations module by responsibility
Moves only: no behavior, policy, message, or logic changes. The
1,494-line automations route file becomes a composition root that mounts
six focused modules in the previous registration order, so route
precedence and both Worker-boundary snapshots are unchanged.

Modules: automation-slack-settings (the two /integration-settings/slack
reads), automation-list, automation-crud (create, get, update, delete),
automation-lifecycle (pause, resume, trigger), automation-runs
(invocations, runs/:runId), automation-keys (regenerate-key). The
validation and target-selection helpers that create and update share
live in automation-validation; the admission constants and the admitted-
automation accessor live in automation-shared.

The 1,895-line test file splits the same way: one suite per module
(create and update separately, so no file passes 1,000 lines), each
dispatching through its own module, with the store doubles, request
builder, sample row, and mock defaults in automations.test-support. All
113 tests kept with their names and intent.
